fix: fixes for social feed cards
