+++ packages/utils/social-feed.ts
@@ -16,8 +16,11 @@ import { getUserId } from "../networks";
 export const DEFAULT_NAME = "Anon";
 export const DEFAULT_USERNAME = "anonymous";
 export const SOCIAL_FEED_ARTICLE_MIN_CHARS_LIMIT = 2500;
+export const SOCIAL_FEED_ARTICLE_TITLE_MAX_CHARS = 145;
+export const SOCIAL_FEED_ARTICLE_SHORT_DESC_MAX_CHARS = 300;
 export const NB_ROWS_SHOWN_IN_PREVIEW = 5;
-export const ARTICLE_COVER_IMAGE_HEIGHT = 240;
+export const ARTICLE_COVER_IMAGE_HEIGHT = 300;
+export const ARTICLE_THUMBNAIL_IMAGE_HEIGHT = 252;
 
 export const getUpdatedReactions = (reactions: Reaction[], icon: string) => {
   const hasIcon = reactions.find((r) => r.icon === icon);
